Allow configuring the background view controller to snapshot
Without this change, presenting from a modally presented view controller will
snapshot the root view which is unexpected(to me at least).

Ideally we could think through this a bit more and have infer the correct
behavior in the most common cases. But this allows someone to manually set
the correct one without breaking existing behavior for anyone.